added functions.py; updated console.py to use new formatting in output console; updated peek.py to clean up and remove functions
